Discover LLM source files across project layouts
