+++ Cargo.toml
@@ -50,6 +50,7 @@ regex = { version = "1", default-features = false, features = ["std", "perf"] }
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls-webpki-roots", "json"] }
 tokio-util = { version = "0.7", default-features = false, features = ["rt"] }
 git2 = { version = "0.20", default-features = false, features = ["https"] }
+glob = { version = "0.3", default-features = false }
 rcgen = { version = "0.13", default-features = false, features = ["crypto", "ring", "pem", "x509-parser"] }
 rusqlite = { version = "0.32", default-features = false, features = ["bundled-sqlcipher-vendored-openssl"] }
 rustls = { version = "0.23", default-features = false, features = ["std", "tls12", "ring", "logging"] }
