+++ crates/loupe-worker/Cargo.toml
@@ -19,6 +19,7 @@ blake3 = { workspace = true }
 base64 = { workspace = true }
 clap = { workspace = true }
 git2 = { workspace = true }
+glob = { workspace = true }
 loupe-core = { path = "../loupe-core" }
 loupe-proto = { path = "../loupe-proto" }
 loupe-tls = { path = "../loupe-tls" }
