+++ crates/loupe-worker/src/scanners/llm_code_review.rs
@@ -18,7 +18,6 @@
 
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use std::time::Duration;
 
 use anyhow::Result;
 use async_trait::async_trait;
@@ -27,141 +26,12 @@ use tokio::sync::Semaphore;
 use tokio_util::sync::CancellationToken;
 
 use crate::llm::prompts::{self, DISCOVERY};
-use crate::llm::{LlmBackend, LlmRequest, DEFAULT_REQUEST_TIMEOUT};
+use crate::llm::{LlmBackend, LlmRequest};
 use crate::scanner::{ScanContext, Scanner};
+use crate::source_discovery::{walk_source_files, ScannerConfig, ScannerConfigPatch};
 
 pub const SCANNER_ID: &str = "llm-code-review";
 const CAPABILITIES: &[&str] = &["scan:llm"];
-
-/// Config knobs operators can set via the repo's `scanner_config`
-/// JSON. Defaults cover Rust, C/C++, JS/TS, Python, Go, Ruby, PHP, and
-/// JVM/Swift sources, plus a broad excludelist that drops common
-/// build / vendor / test dirs across those ecosystems. Tighten or
-/// loosen per-repo by sending a partial JSON override (see
-/// `ScannerConfigPatch`).
-#[derive(Debug, Clone)]
-pub struct ScannerConfig {
-	pub max_concurrent_files: usize,
-	pub max_file_bytes: u64,
-	pub per_request_timeout: Duration,
-	/// File extensions the walk will consider, lower-cased and without
-	/// the leading dot (e.g. `["rs", "cpp"]`).
-	pub include_extensions: Vec<String>,
-	/// Exclusion patterns that disqualify a path. Built-in patterns
-	/// use `dir:<name>` for exact path components and `file:<glob>`
-	/// for basename matches. Legacy custom strings still match as
-	/// path substrings, except `/name` matches an exact component.
-	pub exclude_path_substrings: Vec<String>,
-}
-
-impl Default for ScannerConfig {
-	fn default() -> Self {
-		Self {
-			max_concurrent_files: 8,
-			max_file_bytes: 2 * 1024 * 1024,
-			per_request_timeout: DEFAULT_REQUEST_TIMEOUT,
-			include_extensions: default_extensions(),
-			exclude_path_substrings: default_excludes(),
-		}
-	}
-}
-
-/// Partial override applied on top of `ScannerConfig::default()` (or a
-/// constructor-supplied baseline) when the server passes a non-null
-/// `scanner_config` in the lease envelope. `None` for any field means
-/// "leave the baseline alone"; `Some(...)` replaces the field
-/// wholesale.
-///
-/// Replacing rather than merging is intentional: when an operator
-/// writes `{"include_extensions":["c","h"]}` for a C-only repo they
-/// almost always *don't* want our default Rust/JS/Python/etc.
-/// extensions silently appended.
-#[derive(Debug, Default, Clone, serde::Deserialize)]
-#[serde(default)]
-pub struct ScannerConfigPatch {
-	pub max_concurrent_files: Option<usize>,
-	pub max_file_bytes: Option<u64>,
-	pub per_request_timeout_seconds: Option<u64>,
-	pub include_extensions: Option<Vec<String>>,
-	pub exclude_path_substrings: Option<Vec<String>>,
-}
-
-impl ScannerConfig {
-	pub(crate) fn apply_patch(&mut self, p: ScannerConfigPatch) {
-		if let Some(v) = p.max_concurrent_files {
-			self.max_concurrent_files = v;
-		}
-		if let Some(v) = p.max_file_bytes {
-			self.max_file_bytes = v;
-		}
-		if let Some(v) = p.per_request_timeout_seconds {
-			self.per_request_timeout = Duration::from_secs(v);
-		}
-		if let Some(v) = p.include_extensions {
-			self.include_extensions = v;
-		}
-		if let Some(v) = p.exclude_path_substrings {
-			self.exclude_path_substrings = v;
-		}
-	}
-}
-
-fn default_extensions() -> Vec<String> {
-	[
-		// Rust.
-		"rs", // C / C++ / Obj-C.
-		"c", "h", "cc", "hh", "cpp", "hpp", "cxx", "hxx", "m", "mm", // JS / TS.
-		"js", "jsx", "mjs", "cjs", "ts", "tsx", // Python.
-		"py",  // Go.
-		"go",  // Ruby.
-		"rb",  // PHP.
-		"php", // JVM family.
-		"java", "kt", "kts", "scala", "groovy", // Swift.
-		"swift",  // Misc.
-		"dart", "ex", "exs", "rs.in",
-	]
-	.into_iter()
-	.map(String::from)
-	.collect()
-}
-
-fn default_excludes() -> Vec<String> {
-	[
-		// Rust / Java / general "tests" dirs and matching filename suffixes.
-		"dir:tests",
-		"dir:test",
-		"dir:examples",
-		"dir:__tests__",
-		"file:test_*.*",
-		"file:*_test.*",
-		"file:*_tests.*",
-		"file:*.test.*",
-		"file:*.spec.*",
-		// Build artefacts across ecosystems.
-		"dir:target",
-		"dir:build",
-		"dir:dist",
-		"dir:out",
-		"dir:.next",
-		"dir:.nuxt",
-		"dir:coverage",
-		// Vendored deps.
-		"dir:node_modules",
-		"dir:vendor",
-		"dir:.venv",
-		"dir:venv",
-		"dir:env",
-		// Caches.
-		"dir:__pycache__",
-		"dir:.tox",
-		"dir:.gradle",
-		"dir:.mypy_cache",
-		"dir:.pytest_cache",
-	]
-	.into_iter()
-	.map(String::from)
-	.collect()
-}
 
 pub struct LlmCodeReviewScanner {
 	backend: Arc<dyn LlmBackend>,
@@ -353,205 +223,6 @@ async fn run_one(
 	}
 }
 
-/// Walk the worktree for source files. Strategy:
-/// workspace `[workspace] members` → each member's `src/`;
-/// single-crate `Cargo.toml` → `src/`; otherwise the entire tree
-/// under any `src/` directory. Honours the include extension
-/// allowlist and the path-substring exclude list, plus the per-file
-/// size cap. `.git` and any directory matching an exclude substring
-/// is skipped wholesale.
-pub(crate) fn walk_source_files(workdir: &Path, cfg: &ScannerConfig) -> Vec<PathBuf> {
-	let mut roots: Vec<PathBuf> = Vec::new();
-	let cargo_toml = workdir.join("Cargo.toml");
-	if cargo_toml.exists() {
-		match parse_workspace_members(&cargo_toml) {
-			Some(members) => {
-				for m in members {
-					let p = workdir.join(m).join("src");
-					if p.is_dir() {
-						roots.push(p);
-					}
-				}
-			},
-			None => {
-				let p = workdir.join("src");
-				if p.is_dir() {
-					roots.push(p);
-				}
-			},
-		}
-	}
-	if roots.is_empty() {
-		// Fallback: walk anything under any src/ subdir.
-		roots.push(workdir.to_path_buf());
-	}
-
-	let mut out = Vec::new();
-	for root in roots {
-		walkdir::WalkDir::new(&root)
-			.into_iter()
-			.filter_entry(|e| !is_excluded_dir(e.path(), cfg))
-			.filter_map(|r| r.ok())
-			.filter(|e| e.file_type().is_file())
-			.filter(|e| has_allowed_extension(e.path(), &cfg.include_extensions))
-			.filter(|e| !is_excluded_path(e.path(), &cfg.exclude_path_substrings))
-			.filter(|e| e.metadata().map(|m| m.len() <= cfg.max_file_bytes).unwrap_or(false))
-			.for_each(|e| out.push(e.into_path()));
-	}
-	out.sort();
-	out.dedup();
-	out
-}
-
-/// Lightweight `[workspace] members = [...]` extractor. Returns
-/// `Some(members)` when the manifest has a `[workspace]` section,
-/// `None` otherwise. Deliberately not pulling in `cargo_metadata`
-/// (extra dep, slow); `members` is whatever appears verbatim between
-/// the opening and closing brackets, with each quoted string yielded
-/// in order. Globs like `crates/*` are returned as-is — the caller
-/// only uses each member to build `<workdir>/<member>/src` and skips
-/// non-existent paths.
-fn parse_workspace_members(cargo_toml: &Path) -> Option<Vec<String>> {
-	let text = std::fs::read_to_string(cargo_toml).ok()?;
-	if !text.lines().any(|l| l.trim_start().starts_with("[workspace]")) {
-		return None;
-	}
-
-	// Scan section by section. We only care about the `[workspace]`
-	// section; sections start with `[xxx]` lines.
-	let mut in_workspace = false;
-	let mut buf = String::new(); // buffer for the contents of `members = [ … ]`
-	let mut collecting = false;
-	for line in text.lines() {
-		let trimmed = line.trim();
-		if trimmed.starts_with('[') && trimmed.ends_with(']') {
-			in_workspace = trimmed == "[workspace]";
-			collecting = false;
-			continue;
-		}
-		if !in_workspace {
-			continue;
-		}
-		if collecting {
-			buf.push_str(line);
-			buf.push('\n');
-			if line.contains(']') {
-				collecting = false;
-			}
-			continue;
-		}
-		if let Some(rest) = trimmed.strip_prefix("members") {
-			let rest = rest.trim_start_matches([' ', '\t', '=']).trim();
-			if rest.starts_with('[') {
-				buf.push_str(rest);
-				buf.push('\n');
-				if !rest.contains(']') {
-					collecting = true;
-				}
-			}
-		}
-	}
-
-	// Pull every "..." literal out of the buffer.
-	let mut members = Vec::new();
-	let mut bytes = buf.as_bytes().iter().enumerate();
-	while let Some((i, &b)) = bytes.next() {
-		if b != b'"' {
-			continue;
-		}
-		// Find the closing quote (no escapes — TOML basic strings don't
-		// allow newlines, and we don't expect quotes inside member paths).
-		let rest = &buf.as_bytes()[i + 1..];
-		if let Some(end) = rest.iter().position(|&c| c == b'"') {
-			let s = std::str::from_utf8(&rest[..end]).unwrap_or("");
-			if !s.is_empty() {
-				members.push(s.to_owned());
-			}
-			// Skip past the closing quote in our outer iterator.
-			for _ in 0..end + 1 {
-				bytes.next();
-			}
-		}
-	}
-	Some(members)
-}
-
-fn is_excluded_dir(path: &Path, cfg: &ScannerConfig) -> bool {
-	if has_component(path, ".git") {
-		return true;
-	}
-	cfg.exclude_path_substrings.iter().any(|pattern| matches_exclude(path, pattern))
-}
-
-fn is_excluded_path(path: &Path, excludes: &[String]) -> bool {
-	excludes.iter().any(|pattern| matches_exclude(path, pattern))
-}
-
-fn matches_exclude(path: &Path, pattern: &str) -> bool {
-	if let Some(component) = pattern.strip_prefix("dir:") {
-		return has_component(path, component);
-	}
-	if let Some(glob) = pattern.strip_prefix("file:") {
-		return path
-			.file_name()
-			.and_then(|name| name.to_str())
-			.map(|name| wildcard_match(glob, name))
-			.unwrap_or(false);
-	}
-	if let Some(component) = pattern.strip_prefix('/') {
-		if !component.contains('/') {
-			return has_component(path, component);
-		}
-	}
-	path.to_string_lossy().contains(pattern)
-}
-
-fn has_component(path: &Path, needle: &str) -> bool {
-	path.components().any(|component| component.as_os_str() == needle)
-}
-
-fn wildcard_match(pattern: &str, text: &str) -> bool {
-	let mut remainder = text;
-	let mut parts = pattern.split('*').peekable();
-	let anchored_start = !pattern.starts_with('*');
-	let anchored_end = !pattern.ends_with('*');
-
-	if let Some(first) = parts.next() {
-		if !first.is_empty() {
-			if anchored_start {
-				let Some(rest) = remainder.strip_prefix(first) else { return false };
-				remainder = rest;
-			} else if let Some(pos) = remainder.find(first) {
-				remainder = &remainder[pos + first.len()..];
-			} else {
-				return false;
-			}
-		}
-	}
-
-	let mut last = "";
-	for part in parts {
-		last = part;
-		if part.is_empty() {
-			continue;
-		}
-		if let Some(pos) = remainder.find(part) {
-			remainder = &remainder[pos + part.len()..];
-		} else {
-			return false;
-		}
-	}
-
-	!anchored_end || last.is_empty() || remainder.is_empty()
-}
-
-fn has_allowed_extension(path: &Path, exts: &[String]) -> bool {
-	path.extension()
-		.and_then(|e| e.to_str())
-		.map(|e| exts.iter().any(|allowed| allowed.eq_ignore_ascii_case(e)))
-		.unwrap_or(false)
-}
-
 #[cfg(test)]
 mod tests {
 	use std::sync::atomic::{AtomicUsize, Ordering};
@@ -596,142 +267,6 @@ mod tests {
 			}
 			std::fs::write(p, body).unwrap();
 		}
-	}
-
-	#[test]
-	fn defaults_cover_common_languages() {
-		let cfg = ScannerConfig::default();
-		assert_eq!(cfg.max_file_bytes, 2 * 1024 * 1024);
-		for ext in ["rs", "c", "cpp", "h", "hpp", "js", "ts", "py", "go", "java", "swift"] {
-			assert!(cfg.include_extensions.iter().any(|e| e == ext), "missing: {ext}");
-		}
-		// node_modules and target are excluded out of the box.
-		assert!(cfg.exclude_path_substrings.iter().any(|e| e == "dir:node_modules"));
-		assert!(cfg.exclude_path_substrings.iter().any(|e| e == "dir:target"));
-	}
-
-	#[test]
-	fn patch_overrides_only_the_fields_present() {
-		let mut cfg = ScannerConfig::default();
-		let original_excludes = cfg.exclude_path_substrings.clone();
-		let patch: ScannerConfigPatch =
-			serde_json::from_str(r#"{"include_extensions":["c","h"]}"#).unwrap();
-		cfg.apply_patch(patch);
-		assert_eq!(cfg.include_extensions, vec!["c".to_owned(), "h".to_owned()]);
-		assert_eq!(cfg.exclude_path_substrings, original_excludes);
-	}
-
-	#[test]
-	fn walk_picks_up_non_rust_files_without_cargo_toml() {
-		let tmp = tempfile::tempdir().unwrap();
-		// No Cargo.toml here — fallback walks the workdir.
-		std::fs::create_dir_all(tmp.path().join("src")).unwrap();
-		std::fs::write(tmp.path().join("src/main.cpp"), "// stub\n").unwrap();
-		std::fs::write(tmp.path().join("src/util.h"), "// stub\n").unwrap();
-		std::fs::write(tmp.path().join("src/app.py"), "# stub\n").unwrap();
-		std::fs::write(tmp.path().join("src/page.tsx"), "// stub\n").unwrap();
-
-		let cfg = ScannerConfig::default();
-		let files = walk_source_files(tmp.path(), &cfg);
-		let names: Vec<String> = files
-			.iter()
-			.map(|p| p.strip_prefix(tmp.path()).unwrap().to_string_lossy().into_owned())
-			.collect();
-		for expected in ["src/main.cpp", "src/util.h", "src/app.py", "src/page.tsx"] {
-			assert!(names.iter().any(|n| n == expected), "missing {expected} in {names:?}");
-		}
-	}
-
-	#[test]
-	fn walk_excludes_node_modules_and_build_dirs_by_default() {
-		let tmp = tempfile::tempdir().unwrap();
-		// Real source.
-		std::fs::create_dir_all(tmp.path().join("src")).unwrap();
-		std::fs::write(tmp.path().join("src/index.js"), "// real\n").unwrap();
-		// Vendored deps and build artefacts that must be skipped.
-		std::fs::create_dir_all(tmp.path().join("node_modules/lodash")).unwrap();
-		std::fs::write(tmp.path().join("node_modules/lodash/index.js"), "// vendored\n").unwrap();
-		std::fs::create_dir_all(tmp.path().join("dist")).unwrap();
-		std::fs::write(tmp.path().join("dist/bundle.js"), "// built\n").unwrap();
-
-		let files = walk_source_files(tmp.path(), &ScannerConfig::default());
-		let names: Vec<String> = files
-			.iter()
-			.map(|p| p.strip_prefix(tmp.path()).unwrap().to_string_lossy().into_owned())
-			.collect();
-		assert!(names.iter().any(|n| n == "src/index.js"), "real source missing: {names:?}");
-		assert!(names.iter().all(|n| !n.contains("node_modules")), "leak: {names:?}");
-		assert!(names.iter().all(|n| !n.starts_with("dist/")), "leak: {names:?}");
-	}
-
-	#[test]
-	fn directory_excludes_match_exact_components_not_prefixes() {
-		let tmp = tempfile::tempdir().unwrap();
-		write_crate(
-			tmp.path(),
-			&[
-				("src/outbound_payment.rs", "// production source\n"),
-				("src/out/generated.rs", "// generated output\n"),
-				("src/distill.rs", "// production source\n"),
-				("src/dist/bundle.rs", "// generated output\n"),
-			],
-		);
-
-		let files = walk_source_files(tmp.path(), &ScannerConfig::default());
-		let names: Vec<String> = files
-			.iter()
-			.map(|p| p.strip_prefix(tmp.path()).unwrap().to_string_lossy().into_owned())
-			.collect();
-		assert!(names.iter().any(|n| n == "src/outbound_payment.rs"), "names: {names:?}");
-		assert!(names.iter().any(|n| n == "src/distill.rs"), "names: {names:?}");
-		assert!(names.iter().all(|n| !n.starts_with("src/out/")), "names: {names:?}");
-		assert!(names.iter().all(|n| !n.starts_with("src/dist/")), "names: {names:?}");
-	}
-
-	#[test]
-	fn file_test_patterns_match_common_test_names() {
-		let tmp = tempfile::tempdir().unwrap();
-		write_crate(
-			tmp.path(),
-			&[
-				("src/lib.rs", "// production source\n"),
-				("src/test_utils.rs", "// test helper\n"),
-				("src/accountable_tests.rs", "// test module\n"),
-				("src/payment.test.ts", "// js-style test\n"),
-				("src/payment.spec.ts", "// js-style test\n"),
-			],
-		);
-
-		let files = walk_source_files(tmp.path(), &ScannerConfig::default());
-		let names: Vec<String> = files
-			.iter()
-			.map(|p| p.strip_prefix(tmp.path()).unwrap().to_string_lossy().into_owned())
-			.collect();
-		assert_eq!(names, vec!["src/lib.rs"]);
-	}
-
-	#[test]
-	fn walk_picks_up_src_rs_files_only() {
-		let tmp = tempfile::tempdir().unwrap();
-		write_crate(
-			tmp.path(),
-			&[
-				("src/lib.rs", "// good\n"),
-				("src/util.rs", "// good\n"),
-				("README.md", "ignore\n"),
-				("tests/integration.rs", "// excluded by tests dir\n"),
-			],
-		);
-		let cfg = ScannerConfig::default();
-		let files = walk_source_files(tmp.path(), &cfg);
-		let names: Vec<String> = files
-			.iter()
-			.map(|p| p.strip_prefix(tmp.path()).unwrap().to_string_lossy().into_owned())
-			.collect();
-		assert!(names.iter().any(|n| n.ends_with("src/lib.rs")), "names: {names:?}");
-		assert!(names.iter().any(|n| n.ends_with("src/util.rs")), "names: {names:?}");
-		assert!(names.iter().all(|n| !n.contains("README")), "names: {names:?}");
-		assert!(names.iter().all(|n| !n.contains("tests/")), "names: {names:?}");
 	}
 
 	#[tokio::test]
